crucible-pours-through-shared-insert: a survival right-click with a blob stack on a crucible put the whole stack's goo in the reservoir and consumed one blob. CrucibleInteraction.consumeBlobIntoCrucible inserted BlobStacks.volumeOf(stack), the count times MB_PER_BLOB, then called stack.shrink(1). CrucibleInteraction.tryInsertBlob now pours through the shared BlobInsert.pour with a crucible GooSink. The sink takes the whole units that fit under CrucibleInsertion.reservoirUnitsThatFit (one unit is a blob for a blob stack, one mB for an omniblob), and BlobStacks.deplete shrinks the stack by what was accepted, the way the vat, tap, hub and canister do (decision block-insert-shared). A refused blob still ends the click. CrucibleBasin.holdsNoGoo takes long, matching the long volumes its caller passes.
Red before the fix, runGameTestServer with the new tests against the pre-fix CrucibleInteraction: goo:cr_blob_stack_whole failed "stack spent" (64,000 mB inserted, 63 blobs left in hand), and goo:cr_blob_stack_to_cap read 1,999,997,000 instead of the 2,000,000,000 cap (the whole 64 blob volume refused). Green after: all 174 required gametests pass, cr_cap_blob_in_hand among them (decision diagnose-then-fix-crucible-blob-duplication)

+++ src/main/java/com/mercuriusxeno/goo/block/crucible/CrucibleBasin.java
@@ -147,7 +147,7 @@ public final class CrucibleBasin {
      * @param poolVolume      the goo the partially melted item's pool holds
      * @return true when the reservoir and the pool together hold nothing
      */
-    public static boolean holdsNoGoo(int reservoirVolume, int poolVolume) {
+    public static boolean holdsNoGoo(long reservoirVolume, long poolVolume) {
         return reservoirVolume <= 0 && poolVolume <= 0;
     }
